Replace bin symlink with $PATH append

+++ ellipsis.sh
@@ -24,14 +24,29 @@ pkg.install() {
             wudo $PKG_PATH/choco-install.bash
         fi
     fi
+
+    # Remove old links
+    pkg.unlink
+
+    # Initialize package
+    pkg.init
+}
+
+pkg.init() {
+    # Add package bin to $PATH if it isn't there
+    if [ ! "$(command -v choco)" ]; then
+        export PATH=$PKG_PATH/bin:$PATH
+    fi
 }
 
 pkg.link() {
-    fs.link_rfiles "$PKG_PATH/bin" "$ELLIPSIS_PATH/bin"
+    : # Package does not contain linkable files
 }
 
 pkg.unlink() {
     for file in "$PKG_PATH/bin"/*; do
-        rm "$ELLIPSIS_PATH/bin/$(basename $file)"
+        if [[ -f "$ELLIPSIS_PATH/bin/$(basename $file)" && -L "$ELLIPSIS_PATH/bin/$(basename $file)" ]]; then
+            rm "$ELLIPSIS_PATH/bin/$(basename $file)"
+        fi
     done
 }
